+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Add integration test for sync within one bucket with different encryption
 * Notarize OSX binary
+* File lock arguments and new commands
 
 ### Fixed
 * Fixed breaking integration test case
 * Add zoneinfo to the Windows bundle
 * Fixed unit tests failing on new attributes of FileVersionInfo
+* Removing old buckets in integration tests
 
 ## [2.4.0] - 2021-04-22